# docs/swagger.yaml
definitions:
...
  dozenChairs_internal_models.Image:
    properties:
      created_at:
        type: string
      filename:
        type: string
      id:
        type: string
      product_id:
        type: string
      url:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /api/v1//images/{id}:
    delete:
      parameters:
      - description: Image ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить изображение по ID
      tags:
      - Images
  /api/v1//products/{product_id}/images:
    get:
      parameters:
      - description: ID или slug товара
        in: path
        name: product_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dozenChairs_internal_models.Image'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить изображения по товару
      tags:
      - Images
  /api/v1//upload:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: ID товара
        in: formData
        name: product_id
        required: true
        type: string
      - description: Файлы изображений (можно несколько)
        in: formData
        name: images
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/dozenChairs_internal_models.Image'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Загрузить изображения для продукта
      tags:
      - Images
...
swagger: "2.0"